requirements-dev.txt
-r requirements.txt
django-extensions
django-debug-toolbar
Fabric
flake8
ipython
tox

